Move trace functions to their own module
