+++ scico/util.py
@@ -10,33 +10,17 @@
 
 from __future__ import annotations
 
-import inspect
 import io
 import socket
-import sys
-import types
 import urllib.error as urlerror
 import urllib.request as urlrequest
-import warnings
-from collections import defaultdict
 from functools import reduce, wraps
 from timeit import default_timer as timer
 from typing import Any, Callable, Dict, List, Optional, Sequence, Union
 
-import numpy as np
-
 import jax
 from jax.interpreters.batching import BatchTracer
 from jax.interpreters.partial_eval import DynamicJaxprTracer
-
-from jaxlib.xla_extension import PjitFunction
-
-try:
-    import colorama
-
-    have_colorama = True
-except ImportError:
-    have_colorama = False
 
 
 def rgetattr(obj: object, name: str, default: Optional[Any] = None) -> Any:
@@ -167,217 +151,6 @@ def check_for_tracer(func: Callable) -> Callable:
         return func(*args, **kwargs)
 
     return wrapper
-
-
-def _val_repr(val):
-    if val is None:
-        return "None"
-    elif np.isscalar(val):
-        return str(val)
-    elif isinstance(val, np.dtype):
-        return f"numpy.{val}"
-    elif isinstance(val, tuple) and len(val) < 6:
-        return f"{val}"
-    elif isinstance(val, type):
-        return f"{val.__module__}.{val.__qualname__}"
-    elif isinstance(val, (np.ndarray, jax.Array)):
-        return f"Array{val.shape}"
-    else:
-        if (
-            hasattr(val, "__hash__")
-            and callable(val.__hash__)
-            and val.__hash__() in call_trace.instance_hash
-        ):
-            return f"{call_trace.instance_hash[val.__hash__()]}"
-        else:
-            return f"[{type(val).__name__}]"
-
-
-def call_trace(func: Callable) -> Callable:  # pragma: no cover
-    """Print log of calls to `func`.
-
-    Decorator for printing a log of calls to the wrapped function. A
-    record of call levels is maintained so that call nesting is indicated
-    by call log indentation.
-    """
-    if have_colorama:
-        pre, pst = colorama.Fore.LIGHTRED_EX, colorama.Fore.RESET
-    else:
-        pre, pst = "", ""
-
-    try:
-        method_class = inspect._findclass(func)  # type: ignore
-    except AttributeError:
-        method_class = None
-
-    @wraps(func)
-    def wrapper(*args, **kwargs):
-        name = f"{func.__module__}.{func.__qualname__}"
-        argidx = 0
-        if (
-            args
-            and hasattr(args[0], "__hash__")
-            and callable(args[0].__hash__)
-            and method_class
-            and isinstance(args[0], method_class)
-        ):
-            argidx = 1
-            if args[0].__hash__() in call_trace.instance_hash:
-                name = f"{call_trace.instance_hash[args[0].__hash__()]}.{func.__name__}"
-            elif hasattr(args[0], "__class__"):
-                name = (
-                    f"{args[0].__class__.__module__}.{args[0].__class__.__name__}.{func.__name__}"
-                )
-        argsrep = [_val_repr(val) for val in args[argidx:]]
-        kwargrep = [f"{key}={_val_repr(val)}" for key, val in kwargs.items()]
-        argstr = ", ".join(argsrep + kwargrep)
-        print(
-            f"{pre}>> {' ' * 2 * call_trace.trace_level}{name}({argstr}){pst}",
-            file=sys.stderr,
-        )
-        call_trace.trace_level += 1
-        ret = func(*args, **kwargs)
-        call_trace.trace_level -= 1
-        return ret
-
-    # Set flag indicating that function is already wrapped
-    wrapper._call_trace_wrap = True  # type: ignore
-    # Avoid multiple wrapper layers
-    if hasattr(func, "_call_trace_wrap"):
-        return func
-    else:
-        return wrapper
-
-
-# Call level counter for call_trace decorator
-call_trace.trace_level = 0  # type: ignore
-
-call_trace.instance_hash = {}  # type: ignore
-
-
-def apply_decorator(
-    module: types.ModuleType,
-    decorator: Callable,
-    recursive: bool = True,
-    skip: Optional[Sequence] = None,
-    seen: Optional[defaultdict[str, int]] = None,
-    verbose: bool = False,
-    level: int = 0,
-) -> defaultdict[str, int]:
-    """Apply a decorator function to all functions in a scico module.
-
-    Apply a decorator function to all functions in a scico module,
-    including methods of classes in that module.
-
-    Args:
-        module: The module containing the functions/methods to be
-          decorated.
-        decorator: The decorator function to apply to each module
-          function/method.
-        recursive: Flag indicating whether to recurse into submodules
-          of the specified module. (Hidden modules with a name starting
-          with an underscore are ignored.)
-        skip: A list of class/function/method names to be skipped.
-        seen: A :class:`defaultdict` providing a count of the number of
-          times each function/method was seen.
-        verbose: Flag indicating whether to print a log of functions
-          as they are encountered.
-        level: Counter for recursive call levels.
-
-    Returns:
-        A :class:`defaultdict` providing a count of the number of times
-        each function/method was seen.
-    """
-    indent = " " * 4 * level
-    if skip is None:
-        skip = []
-    if seen is None:
-        seen = defaultdict(int)
-    # Iterate over objects in module
-    for obj_name in dir(module):
-        if obj_name in skip:
-            continue
-        obj = getattr(module, obj_name)
-        if hasattr(obj, "__module__") and obj.__module__[0:5] == "scico":
-            qualname = obj.__module__ + "." + obj.__qualname__
-            if isinstance(obj, (types.FunctionType, PjitFunction)):
-                if not seen[qualname]:  # avoid multiple applications of decorator
-                    setattr(module, obj_name, decorator(obj))
-                seen[qualname] += 1
-                if verbose:
-                    print(f"{indent}Function: {qualname}")
-            elif isinstance(obj, type):
-                if verbose:
-                    print(f"{indent}Class: {qualname}")
-                # Iterate over class attributes
-                for attr_name in dir(obj):
-                    if attr_name in skip:
-                        continue
-                    attr = getattr(obj, attr_name)
-                    if isinstance(attr, (types.FunctionType, PjitFunction)):
-                        qualname = attr.__module__ + "." + attr.__qualname__  # type: ignore
-                        if not seen[qualname]:  # avoid multiple applications of decorator
-                            setattr(obj, attr_name, decorator(attr))
-                        seen[qualname] += 1
-                        if verbose:
-                            print(f"{indent + '    '}Method: {qualname}")
-        elif isinstance(obj, types.ModuleType):
-            if (
-                len(obj.__name__) > len(module.__name__)
-                and obj.__name__[0 : len(module.__name__)] == module.__name__
-            ):
-                short_name = obj.__name__[len(module.__name__) + 1 :]
-            else:
-                short_name = ""
-            if (
-                len(obj.__name__) >= 5
-                and obj.__name__[0:5] == "scico"
-                and len(short_name) > 0
-                and short_name[0] != "_"
-            ):
-                if verbose:
-                    print(f"{indent}Module: {obj.__name__}")
-                if recursive:
-                    seen = apply_decorator(
-                        obj,
-                        decorator,
-                        recursive=recursive,
-                        skip=skip,
-                        seen=seen,
-                        verbose=verbose,
-                        level=level + 1,
-                    )
-    return seen
-
-
-def trace_scico_calls():  # pragma: no cover
-    """Enable tracing of calls to all significant scico functions/methods.
-
-    Enable tracing of calls to all significant scico functions and
-    methods. Note that JIT should be disabled to ensure correct
-    functioning of the tracing mechanism.
-    """
-    if not jax.config.jax_disable_jit:
-        warnings.warn(
-            "Call tracing requested but jit is not disabled. Disable jit"
-            " by setting the environment variable JAX_DISABLE_JIT=1, or use"
-            " jax.config.update('jax_disable_jit', True)."
-        )
-    from scico import (
-        function,
-        functional,
-        linop,
-        loss,
-        metric,
-        operator,
-        optimize,
-        solver,
-    )
-
-    for module in (functional, linop, loss, operator, optimize, function, metric, solver):
-        apply_decorator(module, call_trace, skip=["__repr__"])
-    # Currently unclear why applying this separately is required
-    optimize.Optimizer.solve = call_trace(optimize.Optimizer.solve)
 
 
 def url_get(url: str, maxtry: int = 3, timeout: int = 10) -> io.BytesIO:  # pragma: no cover
